Migrate the media text and post template blocks
