+++ packages/block-library/src/media-text/index.js
@@ -9,6 +9,7 @@ import { privateApis as blocksPrivateApis } from '@wordpress/blocks';
  * Internal dependencies
  */
 import initBlock from '../utils/init-block';
+import { TEMPLATE } from './constants';
 import deprecated from './deprecated';
 import edit from './edit';
 import metadata from './block.json';
@@ -49,6 +50,7 @@ export const settings = {
 		],
 	},
 	transforms,
+	template: TEMPLATE,
 	edit,
 	save,
 	deprecated,
