+++ packages/block-library/src/post-template/edit.js
@@ -21,28 +21,10 @@ import { Spinner, ToolbarGroup } from '@wordpress/components';
 import { store as coreStore } from '@wordpress/core-data';
 import { list, grid } from '@wordpress/icons';
 
-const TEMPLATE = [
-	[ 'core/post-title' ],
-	[
-		'core/post-date',
-		{
-			metadata: {
-				bindings: {
-					datetime: {
-						source: 'core/post-data',
-						args: { field: 'date' },
-					},
-				},
-			},
-		},
-	],
-	[ 'core/post-excerpt' ],
-];
-
 function PostTemplateInnerBlocks( { classList } ) {
 	const innerBlocksProps = useInnerBlocksProps(
 		{ className: clsx( 'wp-block-post', classList ) },
-		{ template: TEMPLATE, __unstableDisableLayoutClassNames: true }
+		{ __unstableDisableLayoutClassNames: true }
 	);
 	return <li { ...innerBlocksProps } />;
 }
